# Protocol speedups