--- a/definitions/client.proto
+++ b/definitions/client.proto
@@ -45,6 +45,7 @@ message Push {
         SUBSCRIBE = 5;
         CONNECT = 6;
         DISCONNECT = 7;
+        REFRESH = 8;
     }
     PushType type = 1;
     string channel = 2;
@@ -103,6 +104,11 @@ message Disconnect {
     uint32 code = 1;
     string reason = 2;
     bool reconnect = 3;
+}
+
+message Refresh {
+    bool expires = 1;
+    uint32 ttl = 2;
 }
 
 message ConnectRequest {