… bench_test.go
… client.pb.go
… clientpb_test.go
… decode_test.go
… definitions/readme.md
… encode.go
… encode_writer.go
… encoding.go
… generate.sh
… go.mod
… go.sum
… raw_test.go
… type.go