-1 alloc for Protobuf unmarshaling and massive speedups for JSON protocol using customized easyjson code generation for encoding and segmentio fast JSON implementation for decoding:

```
❯ benchstat old.txt new.txt
name                               old time/op    new time/op    delta
ReplyProtobufMarshal-12               212ns ± 1%     209ns ± 0%   -1.70%  (p=0.000 n=10+9)
ReplyProtobufMarshalParallel-12       163ns ± 2%     164ns ± 1%     ~     (p=0.388 n=9+10)
ReplyProtobufUnmarshal-12             176ns ± 0%     173ns ± 1%   -1.70%  (p=0.000 n=9+10)
ReplyProtobufUnmarshalParallel-12    50.2ns ±10%    49.6ns ± 6%     ~     (p=0.853 n=10+10)
ReplyJSONMarshal-12                  5.65µs ± 1%    0.47µs ± 0%  -91.67%  (p=0.000 n=9+9)
ReplyJSONMarshalParallel-12          1.12µs ± 8%    0.27µs ± 8%  -75.87%  (p=0.000 n=10+10)
ReplyJSONUnmarshal-12                1.69µs ± 1%    0.46µs ± 0%  -72.99%  (p=0.000 n=9+9)
ReplyJSONUnmarshalParallel-12         535ns ± 8%     122ns ±10%  -77.27%  (p=0.000 n=10+10)

name                               old alloc/op   new alloc/op   delta
ReplyProtobufMarshal-12                864B ± 0%      864B ± 0%     ~     (all equal)
ReplyProtobufMarshalParallel-12        864B ± 0%      864B ± 0%     ~     (all equal)
ReplyProtobufUnmarshal-12              160B ± 0%      128B ± 0%  -20.00%  (p=0.000 n=10+10)
ReplyProtobufUnmarshalParallel-12      160B ± 0%      128B ± 0%  -20.00%  (p=0.000 n=10+10)
ReplyJSONMarshal-12                  1.12kB ± 0%    0.93kB ± 0%  -17.22%  (p=0.000 n=10+10)
ReplyJSONMarshalParallel-12          1.12kB ± 0%    0.93kB ± 0%  -17.22%  (p=0.000 n=10+10)
ReplyJSONUnmarshal-12                1.33kB ± 0%    0.14kB ± 0%  -89.76%  (p=0.000 n=10+10)
ReplyJSONUnmarshalParallel-12        1.33kB ± 0%    0.14kB ± 0%  -89.76%  (p=0.000 n=10+10)

name                               old allocs/op  new allocs/op  delta
ReplyProtobufMarshal-12                3.00 ± 0%      3.00 ± 0%     ~     (all equal)
ReplyProtobufMarshalParallel-12        3.00 ± 0%      3.00 ± 0%     ~     (all equal)
ReplyProtobufUnmarshal-12              5.00 ± 0%      4.00 ± 0%  -20.00%  (p=0.000 n=10+10)
ReplyProtobufUnmarshalParallel-12      5.00 ± 0%      4.00 ± 0%  -20.00%  (p=0.000 n=10+10)
ReplyJSONMarshal-12                    6.00 ± 0%      3.00 ± 0%  -50.00%  (p=0.000 n=10+10)
ReplyJSONMarshalParallel-12            6.00 ± 0%      3.00 ± 0%  -50.00%  (p=0.000 n=10+10)
ReplyJSONUnmarshal-12                  16.0 ± 0%       3.0 ± 0%  -81.25%  (p=0.000 n=10+10)
ReplyJSONUnmarshalParallel-12          16.0 ± 0%       3.0 ± 0%  -81.25%  (p=0.000 n=10+10)
```